remctl: Add a remctl derivation and module
- Basic package installs the binaries
- Module allows installing the daemon, complete with configuring commands,
  installing systemd units and enabling them, etc.

For whatever reason, it's easy to accidentally use two different builds of
remctl -- ensuring that this flake.lock and the microvm one generally used the
same versions seemed to fix it. I was hopeful overlays might let the module use
the package without re-importing all the packages, but it doesn't seem to run.