Fix krb5 dependency
Previously I depended on `krb5` for the Kerberos dependency, which
seemed to work (and still builds on Nixos 24.05). However, on
nixos-unstable, recent changes (perhaps commit 311277204d0) broke the
build. It looks like libkrb5 is the more common dependency to use
(though judging from that commit message, perhaps not the preferred
dependency long-term) and works with all three nixpkg dependencies, so
switch to it.