python-afs: Disable running the checks
It looks like nixos-24.05 tries to run the tests, which expect nose, so
the nixos-24.05 CI build fails now. Unstable nixos doesn't have nose.
Since the tests also expect OpenAFS to be working on the build machine
(they access ThisCell and files in AFS), don't bother fixing the tests
to not use nose, and just stop running the tests during the build.