python-afs: Compatibility with Python 3.13
Python 3.13 introduces a header <lock.h>, which conflicts with OpenAFS's
identically-named header. Python 3.15 will probably rename it to
pylock.h[1], but it's here for Python 3.13 and probably Python 3.14[2].

In the interim, force OpenAFS's headers to come before Python's when
building python-afs. This is a little tricky, because Python's headers
use -I by default while OpenAFS's use -isystem, and -I headers all come
before -isystem ones. However, if we add Python headers with -isystem at
the start of the compiler command, Python's -I will be treated as a dup
of a system header, and Python's headers will be moved to the system
section of the search path. *However*, because they're at the *start* of
the search path, they'll still beat OpenAFS, so we need to also add
-isystem for OpenAFS at the start of the compiler command, so the
relative order among the system section is correct.

Upstream mit-athena/python-afs has an issue for this[3], but because Nix
builds weirdly, the patch I wrote there doesn't work for Nix.

[1] https://github.com/python/cpython/issues/136759
[2] https://github.com/python/cpython/pull/137075#issuecomment-3126856020
[3] https://github.com/mit-athena/python-afs/issues/6