discuss: Use new e2fsprogs.scripts output
In nixpkgs commit e58c9e7dd621a, mk_cmds got split off from the default
e2fsprogs output into a new e2fsprogs.scripts output, which means we
need to pull in that output to build. It's added to nativeBuildInputs
because it's a script that gets run at build time, whereas AIUI
buildInputs is for things like libraries that get linked in at build
time but used at run time (which is relevant for cross-compiling).

While I'm at it, update flake.lock to pull in a newer nixpkgs version
that actually had this problem.